feat: add utils folder with wsResponses and validators, extract Google login to googleService
