+++ app backend/services/googleService.js
@@ -1,0 +1,20 @@
+// services/googleService.js
+const { OAuth2Client } = require('google-auth-library');
+const client = new OAuth2Client(process.env.GOOGLE_CLIENT_ID);
+
+async function verifyGoogleToken(token) {
+  const ticket = await client.verifyIdToken({
+    idToken: token,
+    audience: process.env.GOOGLE_CLIENT_ID,
+  });
+
+  const payload = ticket.getPayload();
+  return {
+    email: payload.email,
+    name: payload.name,
+  };
+}
+
+module.exports = {
+  verifyGoogleToken
+};
