+++ app backend/utils/validators.js
@@ -1,0 +1,10 @@
+function sendSuccess(ws, type, payload) {
+    ws.send(JSON.stringify({ type, ...payload }));
+}
+  
+function sendError(ws, type, reason) {
+    ws.send(JSON.stringify({ type, reason }));
+}  
+  
+module.exports = { sendSuccess, sendError };
+
